Refactor poll package
