+++ Core/Poll/PollHospitalSQL.cs
@@ -1,5 +1,4 @@
 ﻿using HealthCareInfromationSystem.models.entity;
-using HealthCareInfromationSystem.repository.users;
 using HealthCareInfromationSystem.utils;
 using System;
 using System.Collections.Generic;
@@ -8,48 +7,50 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HealthCareInfromationSystem.repository
+namespace HealthCareInfromationSystem.Core.Poll
 {
-    class PollDoctorSQL : IPollDoctorRepo
+    class PollHospitalSQL : IPollHospitalRepo
     {
-        public void Add(PollDoctor pollDoctor)
+        public void Add(PollHospital pollHospital)
         {
             using (OleDbConnection connection = new OleDbConnection(Constants.connectionString))
             {
                 connection.Open();
-                string query = $"insert into poll_doctor values (\"{pollDoctor.Id}\", \"{pollDoctor.Quality}\", \"{pollDoctor.Recommendation}\"," +
-                    $" \"{pollDoctor.Comment}\", \"{pollDoctor.Doctor.Id}\", \"{pollDoctor.Patient.Id}\")";
+                string query = $"insert into poll_hospital values (\"{pollHospital.Id}\", \"{pollHospital.Quality}\", " +
+                    $"\"{pollHospital.Cleanliness}\", \"{pollHospital.Impression}\", \"{pollHospital.Recommendation}\"," +
+                    $" \"{pollHospital.Comment}\", \"{pollHospital.Patient.Id}\")";
                 OleDbCommand command = new OleDbCommand(query, connection);
                 command.ExecuteNonQuery();
             }
         }
 
-        public void Edit(PollDoctor pollDoctor)
+        public void Edit(PollHospital pollHospital)
         {
             using (OleDbConnection connection = new OleDbConnection(Constants.connectionString))
             {
                 connection.Open();
-                string query = $"update poll_doctor set quality=\"{pollDoctor.Quality}\", recommendation=\"{pollDoctor.Recommendation}\", " +
-                    $"comment=\"{pollDoctor.Comment}\", doctorId=\"{pollDoctor.Doctor.Id}\", patientId=\"{pollDoctor.Patient.Id}\" " +
-                    $"where ID=\"{pollDoctor.Id}\"";
+                string query = $"update poll_hospital set quality=\"{pollHospital.Quality}\", cleanliness=\"{pollHospital.Cleanliness}\", " +
+                    $"impression=\"{pollHospital.Impression}\", recommendation=\"{pollHospital.Recommendation}\", " +
+                    $"comment=\"{pollHospital.Comment}\", patientId=\"{pollHospital.Patient.Id}\" " +
+                    $"where ID=\"{pollHospital.Id}\"";
                 OleDbCommand command = new OleDbCommand(query, connection);
                 command.ExecuteNonQuery();
             }
         }
 
-        public List<PollDoctor> LoadAll()
+        public List<PollHospital> LoadAll()
         {
             using (OleDbConnection connection = new OleDbConnection(Constants.connectionString))
             {
-                OleDbCommand command = new OleDbCommand("select * from poll_doctor", connection);
+                OleDbCommand command = new OleDbCommand("select * from poll_hospital", connection);
 
                 connection.Open();
                 OleDbDataReader reader = command.ExecuteReader();
-                List<PollDoctor> polls = new List<PollDoctor>();
+                List<PollHospital> polls = new List<PollHospital>();
 
                 while (reader.Read())
                 {
-                    PollDoctor poll = PollDoctor.Parse(reader);
+                    PollHospital poll = PollHospital.Parse(reader);
                     polls.Add(poll);
                 }
                 reader.Close();
