+++ Core/Poll/PollService.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HealthCareInfromationSystem.Servise
+namespace HealthCareInfromationSystem.Core.Poll
 {
     class PollService
     {
